Move Save(IObject3D, path) to Object3D, supports stl, amf, obj, mcx?
