+++ DataConverters3D/MeshFileIo.cs
@@ -42,69 +42,6 @@ namespace MatterHackers.DataConverters3D
 {
 	public static class MeshFileIo
 	{
-		public static bool Save(IObject3D item, string meshPathAndFileName, CancellationToken cancellationToken, MeshOutputSettings outputInfo = null, Action<double, string> reportProgress = null)
-		{
-			try
-			{
-				if (outputInfo == null)
-				{
-					outputInfo = new MeshOutputSettings();
-				}
-				switch (Path.GetExtension(meshPathAndFileName).ToUpper())
-				{
-					case ".STL":
-						Mesh mesh = DoMergeAndTransform(item, outputInfo, cancellationToken);
-						return StlProcessing.Save(mesh, meshPathAndFileName, cancellationToken, outputInfo);
-
-					case ".AMF":
-						outputInfo.ReportProgress = reportProgress;
-						return AmfDocument.Save(item, meshPathAndFileName, outputInfo);
-
-					case ".OBJ":
-						outputInfo.ReportProgress = reportProgress;
-						return ObjSupport.Save(item, meshPathAndFileName, outputInfo);
-
-					default:
-						return false;
-				}
-			}
-			catch (Exception)
-			{
-				return false;
-			}
-		}
-
-		public static Mesh DoMergeAndTransform(IObject3D item, MeshOutputSettings outputInfo, CancellationToken cancellationToken)
-		{
-			var visibleMeshes = item.VisibleMeshes().Where((i) => i.WorldPersistable());
-			if (visibleMeshes.Count() == 1)
-			{
-				var first = visibleMeshes.First();
-				if(first.WorldMatrix() == Matrix4X4.Identity)
-				{
-					return first.Mesh;
-				}
-			}
-
-			Mesh allPolygons = new Mesh();
-
-			foreach (var rawItem in visibleMeshes)
-			{
-				var mesh = rawItem.Mesh.Copy(cancellationToken);
-				mesh.Transform(rawItem.WorldMatrix());
-				if (outputInfo.CsgOptionState == MeshOutputSettings.CsgOption.DoCsgMerge)
-				{
-					allPolygons = CsgOperations.Union(allPolygons, mesh, null, cancellationToken);
-				}
-				else
-				{
-					allPolygons.CopyFaces(mesh);
-				}
-			}
-
-			return allPolygons;
-		}
-
 		public static long GetEstimatedMemoryUse(string fileLocation)
 		{
 			switch (Path.GetExtension(fileLocation).ToUpper())
